requirements.txt
aiogram==2.25.1
python-dotenv
aiohttp
apscheduler
Flask==2.2.5
Flask-Admin==1.6.1
Flask-SQLAlchemy
psycopg2-binary	
aiogram<3.0.0
